+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+# 0.4.0 (released on 2020-01-25)
+
+- Add --quiet flag to mute all stdout output.
+- Add --verbose flag to print all incoming connections and what their queried host.
 
 # 0.3.0 (released on 2020-01-24)
 
-- Switch from Rocket to actix-web to make this work on stable Rust
+- Switch from Rocket to actix-web to make this work on stable Rust.